Fixes NPE when building from URI
According to the URI API, both the path and query getters can return
null. While that isn't usually the case, this test covers those
scenarios. The resulting URL may or may not be the desired outcome: for
example, any mailto URI will result in the URL "mailto:". I believe this
would also be the case for other SSP-style URIs.